# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IPAM Next API
  description: Route-complete REST API specification for the current backend.
  version: 1.13.0
  contact:
    name: IPAM Next
servers:
  - url: /
    description: Same-origin API
security:
  - sessionCookie: []
  - bearerToken: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerToken:
      type: http
      scheme: bearer
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  schemas:
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        mfa_required:
          type: boolean
        mfa_token:
          type: string
...
  requestBodies:
    JSON:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenericObject"
...
  responses:
    OK:
      description: Successful response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenericObject"
...
    Created:
      description: Resource created
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GenericObject"
...
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in with username and password
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          description: Login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a local user account
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify an email address
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/request-password-reset:
    post:
      tags: [Auth]
      summary: Request password reset
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/unlock:
    get:
      tags: [Security]
      summary: Verify account unlock token
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/OK"
    post:
      tags: [Security]
      summary: Request account unlock
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/verify-mfa:
    post:
      tags: [Auth]
      summary: Complete MFA verification
      security: []
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /api/v1/auth/me/tokens:
    get:
      tags: [Auth]
      summary: List personal API tokens
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [Auth]
      summary: Create personal API token
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/auth/me/tokens/{id}/rotate:
    post:
      tags: [Auth]
      summary: Rotate personal API token
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/me/tokens/{id}/extend:
    post:
      tags: [Auth]
      summary: Extend personal API token expiration
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /api/v1/auth/me/sessions/{sessionID}:
    delete:
      tags: [Auth]
      summary: Revoke one session
      parameters:
        - $ref: "#/components/parameters/sessionID"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/me/mfa:
    get:
      tags: [Auth]
      summary: Get MFA status
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Auth]
      summary: Disable TOTP MFA
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /api/v1/auth/me/mfa/confirm:
    post:
      tags: [Auth]
      summary: Confirm TOTP MFA setup
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/auth/me/mfa/backup-codes:
    post:
      tags: [Auth]
      summary: Regenerate MFA backup codes
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
  /api/v1/user/notification-preferences:
    get:
      tags: [Users]
      summary: Get notification preferences
      responses:
        "200":
          $ref: "#/components/responses/OK"
    put:
      tags: [Users]
      summary: Update notification preferences
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
  /api/v1/users/{id}/role:
    put:
      tags: [Users]
      summary: Update user role
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/sections:
    get:
      tags: [Sections]
      summary: List sections
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [Sections]
      summary: Create section
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/sections/search:
    post:
      tags: [Sections]
      summary: Search sections
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
  /api/v1/sections/{id}:
    get:
      tags: [Sections]
      summary: Get section
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    put:
      tags: [Sections]
      summary: Update section
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Sections]
      summary: Delete section
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
  /api/v1/sections/{id}/subnets/tree:
    get:
      tags: [Sections, Subnets]
      summary: Get subnet tree for a section
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
  /api/v1/sections/{id}/topology:
    get:
      tags: [Sections]
      summary: Get section topology
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/sections/{sectionID}/subnets:
    get:
      tags: [Subnets]
      summary: List section subnets
      parameters:
        - $ref: "#/components/parameters/sectionID"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [Subnets]
      summary: Create subnet
      parameters:
        - $ref: "#/components/parameters/sectionID"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
...
  /api/v1/subnets/{id}:
    get:
      tags: [Subnets]
      summary: Get subnet
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    put:
      tags: [Subnets]
      summary: Update subnet
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Subnets]
      summary: Delete subnet
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
  /api/v1/subnets/{subnetID}/utilization:
    get:
      tags: [Subnets]
      summary: Get subnet utilization
      parameters:
        - $ref: "#/components/parameters/subnetID"
      responses:
        "200":
          $ref: "#/components/responses/OK"
  /api/v1/subnets/search/{sectionID}:
    post:
      tags: [Subnets]
      summary: Search subnets in a section
      parameters:
        - $ref: "#/components/parameters/sectionID"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
  /api/v1/subnets/{subnetID}/ip-addresses:
    get:
      tags: [IP Addresses]
      summary: List IP addresses
      parameters:
        - $ref: "#/components/parameters/subnetID"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [IP Addresses]
      summary: Create IP address
      parameters:
        - $ref: "#/components/parameters/subnetID"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/subnets/{subnetID}/ip-addresses/allocate:
    post:
      tags: [IP Addresses]
      summary: Allocate next available IP address
      parameters:
        - $ref: "#/components/parameters/subnetID"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/subnets/{id}/scan-results:
    get:
      tags: [Subnets, Discovery]
      summary: Get subnet scan results
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
  /api/v1/subnets/{id}/delegations:
    get:
      tags: [Subnets]
      summary: List IPv6 delegations
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
    post:
      tags: [Subnets]
      summary: Create IPv6 delegation
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "201":
          $ref: "#/components/responses/Created"
  /api/v1/subnets/{id}/utilisation/history:
    get:
      tags: [Subnets, Reports]
      summary: Get subnet utilisation history
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OKArray"
  /api/v1/delegations/{id}:
    put:
      tags: [Subnets]
      summary: Update IPv6 delegation
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        $ref: "#/components/requestBodies/JSON"
      responses:
        "200":
          $ref: "#/components/responses/OK"
    delete:
      tags: [Subnets]
      summary: Delete IPv6 delegation
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
...
  /api/v1/ip-addresses/{id}/release:
    post:
      tags: [IP Addresses]
      summary: Release IP address
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          $ref: "#/components/responses/OK"
...